Renamed and fixed typo.
