+++ src/components/SelectContestant.tsx
